+++ popup-bom/popup-parent/popup-component/popup-common/src/main/java/com/photowey/component/common/callback/BiCallback.java
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.photowey.component.common.fx;
+package com.photowey.component.common.callback;
 
-import java.util.function.Consumer;
+import java.util.function.BiConsumer;
 
 /**
- * {@code SuccessCallback}
+ * {@code BiCallback}
  *
  * @author photowey
  * @date 2022/12/25
  * @since 1.0.0
  */
 @FunctionalInterface
-public interface SuccessCallback<T> extends Consumer<T> {
+public interface BiCallback<T, U> extends BiConsumer<T, U> {
 }
