+++ popup-bom/popup-parent/popup-component/popup-common/src/main/java/com/photowey/component/common/callback/FailureCallback.java
