+++ popup-bom/popup-parent/popup-starter/popup-starter-validator/README.md
@@ -1,0 +1,3 @@
+# `popup-starter-validator`
+
+> `Popup` validator starter module
