+++ popup-bom/popup-parent/popup-starter/popup-starter-validator/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.photowey.popup.starter.validator.config.PopupValidatorAutoConfigure
